# config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "GREEN AGRIC API",
    "version": "1.0.0",
    "description": "Backend API & MQTT Service cho hệ thống nông nghiệp thông minh"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    { "bearerAuth": [] }
  ],
  "paths": {
...
    "/sensor/{id}": {
      "put": {
        "summary": "Cập nhật cảm biến",
        "tags": ["Sensor"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Updated" } }
      },
      "delete": {
        "summary": "Xóa cảm biến",
        "tags": ["Sensor"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Deleted" } }
      }
    },
    "/sensor/area/{area_id}/latest": {
      "get": {
        "summary": "Dữ liệu cảm biến mới nhất theo khu vực",
        "tags": ["Sensor"],
        "parameters": [
          { "name": "area_id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/sensor/area/{area_id}/history/{type}": {
      "get": {
        "summary": "Lịch sử cảm biến theo loại",
        "tags": ["Sensor"],
        "parameters": [
          { "name": "area_id", "in": "path", "required": true, "schema": { "type": "integer" } },
          { "name": "type", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
...
    "/device/{id}": {
      "put": {
        "summary": "Cập nhật thiết bị",
        "tags": ["Device"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Updated" } }
      },
      "delete": {
        "summary": "Xóa thiết bị",
        "tags": ["Device"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Deleted" } }
      }
    },
...
  }
}